requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.2.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
numpy==1.26.0